- Display error when Port is already in use. - Change os exit status. - checkVersionCloudSqlProxy use when init. - disable display cloud-sql-proxy version.
