+++ cmd/doctor.go
@@ -87,24 +87,3 @@ func doctor() {
 		_, _ = boldGreen.Println("Your system is All Green!")
 	}
 }
-
-func checkVersionCloudSqlProxy() {
-	red := color.New(color.FgRed)
-	boldRed := red.Add(color.Bold)
-	green := color.New(color.FgGreen)
-
-	boldGreen := green.Add(color.Bold)
-
-	// Check Version cloud-sql-proxy
-	cloudsqlproxyversion := exec.Command("cloud-sql-proxy", "--version")
-	cloudsqlproxyversionout, err := cloudsqlproxyversion.Output()
-	if err != nil {
-		_, _ = boldRed.Println("Error: %s", err)
-		_, _ = boldRed.Println("Please upgrade your cloud-sql-proxy version to 2 or higher")
-		_, _ = boldGreen.Println("Install URL:https://cloud.google.com/sql/docs/postgres/sql-proxy?hl=ja#install")
-		os.Exit(1)
-
-	} else {
-		fmt.Printf("cloud-sql-proxy version: %s", cloudsqlproxyversionout)
-	}
-}
