+++ cmd/list.go
@@ -8,7 +8,6 @@ import (
 )
 
 func listInstance() []string {
-	checkVersionCloudSqlProxy()
 	command := fmt.Sprintf("ps aux | grep cloud-sql-proxy | grep -v grep | grep -v 'awk -F cloud-sql-proxy' | awk -F 'cloud-sql-proxy ' '{print $NF}'")
 	processlist := exec.Command("bash", "-c", command)
 	output, _ := processlist.Output()
@@ -16,7 +15,7 @@ func listInstance() []string {
 	list := strings.Split(line, "\n")
 	if list[0] == "" {
 		fmt.Println("No Instance connected")
-		os.Exit(1)
+		os.Exit(0)
 	}
 
 	return list
